Breaking out Menus
The Menu screens were getting a little wild. So I broke up them up so we now have Main menu, settings menu, and file manager.
