+++ .gitignore
@@ -1,3 +1,5 @@
 ./Old_Versions
 freewrites
+TypeWrytes
 __pycache__
+client_secret.json
